fix: 编辑动图初始化渲染bug
