+++ packages/web/CHANGELOG.md
@@ -1,6 +1,10 @@
 # @pear-rec/web
 
-## 1.3.10
+## 1.3.12
+
+fix: 编辑动图初始化渲染bug
+
+## 1.3.11
 
 perf: 录制动图优化
