+++ packages/web/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@pear-rec/web",
   "private": true,
-  "version": "1.3.10",
+  "version": "1.3.12",
   "scripts": {
     "dev": "vite",
     "build": "rimraf dist && tsc && vite build",
