+++ packages/web/src/pages/recorderScreen/index.tsx
@@ -1,8 +1,6 @@
 import React, { useState, useEffect, useRef } from 'react';
 import { useTranslation } from 'react-i18next';
 import ininitApp from '../../pages/main';
-import { useApi } from '../../api';
-import { useUserApi } from '../../api/user';
 import SelectMedia from '../../components/recorderScreen/SelectMedia';
 import CropRecorder from '../../components/recorderScreen/CropRecorder';
 import ScreenRecorder from '../../components/recorderScreen/ScreenRecorder';
@@ -10,27 +8,10 @@ import '@pear-rec/timer/src/Timer/index.module.scss';
 import styles from './index.module.scss';
 
 const RecorderScreen = () => {
-  const userApi = useUserApi();
-  const userRef = useRef({} as any);
   const videoRef = useRef(null);
   const iframeRef = useRef(null);
   const [isIframe, setIsIframe] = useState(true);
   const [isMedia, setIsMedia] = useState(false);
-
-  useEffect(() => {
-    window.isOffline || userRef.current.id || getCurrentUser();
-  }, []);
-
-  async function getCurrentUser() {
-    try {
-      const res = (await userApi.getCurrentUser()) as any;
-      if (res.code == 0) {
-        userRef.current = res.data;
-      }
-    } catch (err) {
-      console.log(err);
-    }
-  }
 
   function setMediaStream(mediaStream) {
     videoRef.current!.srcObject = mediaStream;
